+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add default description for relationship serializer submatcher
+
 ## [1.1.0] - 2021-04-15
 
 - Add `.serializer` submatcher for relationship matchers (`belong_to`, `have_one` and `have_many`)